sphere stage 2 working with normals

+++ vec3.h
@@ -8,24 +8,24 @@ class vec3
 {
     public:
 		  vec3(){}
-	 	  vec3(double i, double j, double k)
+	 	  vec3(float i, float j, float k)
 		  {
 	 		  this->i = i;
 		  	  this->j = j;
 			  this->k = k;
 		  }
 
-		  inline double x () const {return i;}
-		  inline double y () const {return j;}
-		  inline double z () const {return k;}
-		  inline double r () const {return i;}
-		  inline double g () const {return j;}
-		  inline double b () const {return k;}
+		  inline float x () const {return i;}
+		  inline float y () const {return j;}
+		  inline float z () const {return k;}
+		  inline float r () const {return i;}
+		  inline float g () const {return j;}
+		  inline float b () const {return k;}
 		  
 		  inline const vec3& operator + () const {return *this;}
 		  inline vec3 operator - () const {return vec3(-i, -j, -k);}
 		  
-		  inline double operator [] (int i) const 
+		  inline float operator [] (int i) const 
 		  {	
 				if (i == 0)
 				{
@@ -47,7 +47,7 @@ class vec3
 				}
 		  }
 		   
-		  inline double& operator [] (int i) 
+		  inline float& operator [] (int i) 
 		  {
 				if (i == 0)
 				{
@@ -72,22 +72,22 @@ class vec3
 		  inline vec3& operator -= (const vec3 &v2);
 		  inline vec3& operator *= (const vec3 &v2);
 		  inline vec3& operator /= (const vec3 &v2);
-		  inline vec3& operator *= (const double d);
-		  inline vec3& operator /= (const double d);
+		  inline vec3& operator *= (const float d);
+		  inline vec3& operator /= (const float d);
 		  
-		  inline double length() const
+		  inline float length() const
 		  {
 				return sqrt(this->i*this->i + this->j*this->j + this->k*this->k);
 		  }
 
-		  inline double squared_length() const
+		  inline float squared_length() const
 		  {
 				return this->i*this->i + this->j*this->j + this->k*this->k;
 		  }
 
 		  inline void normalize();
 		  
-		  double i, j, k;
+		  float i, j, k;
 };
 
 inline ostream& operator << (ostream &out, const vec3 &t)
@@ -122,22 +122,22 @@ inline vec3 operator / (const vec3 &v1, const vec3 &v2)
 	 return vec3(v1.i/v2.i, v1.j/v2.j, v1.k/v2.k);
 }
 
-inline vec3 operator * (double d, const vec3 &v)
+inline vec3 operator * (float d, const vec3 &v)
 {
 	 return vec3(d*v.i, d*v.j, d*v.k);
 }
 
-inline vec3 operator / (const vec3 &v, double d)
+inline vec3 operator / (const vec3 &v, float d)
 {
 	 return vec3(v.i/d, v.j/d, v.k/d);
 }
 
-inline vec3 operator * (const vec3 &v, double d)
+inline vec3 operator * (const vec3 &v, float d)
 {
 	 return vec3(v.i*d, v.j*d, v.k*d);
 }
 
-inline double dot(const vec3 &v1, const vec3 &v2)
+inline float dot(const vec3 &v1, const vec3 &v2)
 {
 	 return (v1.i*v2.i + v1.j*v2.j + v1.k*v2.k);
 }
@@ -179,7 +179,7 @@ inline vec3& vec3::operator /= (const vec3 &v)
 	 return *this;
 }
 
-inline vec3& vec3::operator *= (const double d)
+inline vec3& vec3::operator *= (const float d)
 {
 	 this->i *= d;
 	 this->j *= d;
@@ -187,7 +187,7 @@ inline vec3& vec3::operator *= (const double d)
 	 return *this;
 }
 
-inline vec3& vec3::operator /= (const double d)
+inline vec3& vec3::operator /= (const float d)
 {
 	 this->i /= d;
 	 this->j /= d;
@@ -213,7 +213,7 @@ int main()
 	 cout << "ostream : " << v1 << "\n" << endl;
 	 // test for >>
 	 vec3 v3;
-	 double arr[3];
+	 float arr[3];
 	 cout << "enter 3 numbers and be sure to hit <Enter> after each one" << endl;
 	 cin >> v3;
 	 cout << "the vec3 you created is v3 : [" << v3.i << ", " << v3.j << ", " << v3.k << "]\n" << endl;
@@ -246,7 +246,7 @@ int main()
 	 v9 *= v1;
 	 cout << "v9*=v1 == [" << v9.i << ", " << v9.j << ", " << v9.k << "]\n" << endl;
     
-	 double v10 = dot(v1,v1);
+	 float v10 = dot(v1,v1);
 	 cout << "v10 = dot(v1,v1) == " << v10 << endl;
 
 	 vec3 v11 = cross(v1,v1);
